requirements.txt
# Kafka + Spark Streaming + Delta Lake Project
pyspark==3.5.0
kafka-python==2.0.2
delta-spark==3.2.0

# Visualization & Dashboard
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3
